# core/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        },
                        "timestamp": {
                          "type": "string",
                          "example": "2026-01-01T00:00:00.000Z"
                        },
                        "data": {
                          "$ref": "#/components/schemas/AuthResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh access token using refresh token",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "nestjs-starter",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    },
    {
      "url": "https://api.staging.example.com",
      "description": "Staging"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": ["accessToken", "refreshToken"]
      },
...
    }
  }
}